Replace PulsarController by PulsarConnector
