+++ core/src/main/scala/com/raphtory/client/QuerySender.scala
@@ -3,7 +3,6 @@ package com.raphtory.client
 import com.raphtory.components.querymanager.Query
 import com.raphtory.components.querytracker.QueryProgressTracker
 import com.raphtory.config.ComponentFactory
-import com.raphtory.config.PulsarController
 import com.raphtory.config.Scheduler
 import com.raphtory.config.TopicRepository
 import com.raphtory.serialisers.PulsarKryoSerialiser
