+++ core/src/main/scala/com/raphtory/client/RaphtoryClient.scala
@@ -8,7 +8,6 @@ import com.raphtory.components.Component
 import com.raphtory.components.querymanager.Query
 import com.raphtory.components.querytracker.QueryProgressTracker
 import com.raphtory.config.ComponentFactory
-import com.raphtory.config.PulsarController
 import com.raphtory.serialisers.PulsarKryoSerialiser
 import com.typesafe.config.Config
 import com.typesafe.scalalogging.Logger
