+++ core/src/main/scala/com/raphtory/components/querymanager/QueryManager.scala
@@ -1,7 +1,6 @@
 package com.raphtory.components.querymanager
 
 import com.raphtory.components.Component
-import com.raphtory.config.PulsarController
 import com.raphtory.config.Scheduler
 import com.raphtory.config.TopicRepository
 import com.typesafe.config.Config
